// src/api/v2/suppliers/suppliers.types.ts
export type TCreateSupplierClothItem = {
  code: string;
  name?: string;
  cloth_type_id?: number;
  category_id?: number;
  subcategory_id?: number;
  entity_type: "branch" | "factory" | "workshop";
  entity_id: number;
  price: number;
};

export type TCreateSupplierRequest = {
  name: string;
  code: string;
  phone?: string;
  address?: string;
  category_id: number;
  subcategory_id: number;
  branch_id: number;
  order_date: string;
  total_amount: number;
  payment_amount: number;
  remaining_payment: number;
  notes?: string;
  clothes: TCreateSupplierClothItem[];
};